pyproject.toml
[project]
name = "genai-agent"
version = "0.1.0"
description = "FastAPI application for downloading files and summarizing PDFs using Google Gemini"
authors = [
    {name = "Ricardo Dinis", email = "basalto@gmail.com"}
]
dependencies = [
    "fastapi==0.109.0",
    "uvicorn==0.27.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "PyPDF2==3.0.1",
    "google-generativeai==0.3.2",
    "python-dotenv==1.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["main.py"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx==0.24.1",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]
